+++ examples/app-router/public/favicon.ico
