+++ examples/app-router/src/app/forced-color-scheme/[colorScheme]/page.tsx
