+++ examples/app-router/src/app/forced-color-scheme/page.tsx
