+++ examples/app-router/src/app/page.tsx
@@ -1,0 +1,8 @@
+export const metadata = {
+  title: "Nextjs Themes",
+};
+
+/** next.js landing page */
+export default function Page(): JSX.Element {
+  return <></>;
+}
