+++ examples/app-router/src/app/themed-page/[theme]/page.tsx
