+++ examples/app-router/src/app/themed-page/page.tsx
