+++ packages/shared/src/global.scss
@@ -151,6 +151,7 @@ a {
   margin: 20px auto;
   display: block;
   max-width: 1100px;
+  text-align: center;
 }
 
 @media screen and (min-width: 800px) {
